Add Repeating.has_children() method
