+++ stere/strategy/splinter.py
@@ -6,26 +6,7 @@ from stere import Stere
 
 from .strategy import strategy
 
-
-def _retry(fn, retry_time: typing.Optional[int] = None) -> bool:
-    """Retry a function for a specific amount of time.
-
-    Returns:
-        True if the function returns a truthy value, else False
-
-    Arguments:
-        fn: Function to retry
-        retry_time: Number of seconds to retry. If not specified,
-            Stere.retry_time will be used.
-
-    """
-    retry_time = retry_time or Stere.retry_time
-    end_time = time.time() + retry_time
-
-    while time.time() < end_time:
-        if fn():
-            return True
-    return False
+from ..utils import _retry
 
 
 class SplinterBase:
