+++ tests/splinter/test_field.py
@@ -6,18 +6,18 @@ import pytest
 from selenium.webdriver.remote.remote_connection import LOGGER
 
 from stere.fields import Field
-from stere.fields.field import _try_until_timeout
+from stere.fields.field import _retry
 
 LOGGER.setLevel(logging.WARNING)
 
 
-def test_try_until_timeout():
-    """When I call _try_until_timeout
+def test_retry():
+    """When I call _retry
     Then a function is called until it returns a True value
     """
     now = time.time()
 
-    result = _try_until_timeout(
+    result = _retry(
         func=lambda: True if time.time() >= (now + 6) else False,
         wait_time=8,
     )
@@ -25,14 +25,14 @@ def test_try_until_timeout():
     assert result
 
 
-def test_try_until_timeout_fails():
-    """When I call _try_until_timeout
+def test_retry_fails():
+    """When I call _retry
     And the timeout is hit
     Then it returns False
     """
     now = time.time()
 
-    result = _try_until_timeout(
+    result = _retry(
         func=lambda: True if time.time() == (now + 6) else False,
         wait_time=4,
     )
